Resolve Ruby warning by side stepping forwardable
Ruby produces a warning when it uses the forwardable module to forward a
method to a class that utilises `method_missing` to define methods.

Since we only have one method that's forwarded it's simpler to just add
a method ourselves than utilise Forwardable.

It looks like this lingers somewhat unresolved in Ruby land: https://bugs.ruby-lang.org/issues/13326