Do not require requested_user to respond to #job
Zendesk itself doesn't require the `details` attribute to be set and,
looking in the git history for both this repo and the Support app repo
(where this code seems to have originated), I can't find any rationale
for why the current code expects a job title to be provided.

The only use of `Users#create_or_update_user` [1] seems to be in the
Support app. And in this pull request [2] we changed the relevant form
so that it no longer collects the requested user's job title. So it
seems to make sense to remove the requirement for the required user to
respond to the `#job` method.

[1]: https://github.com/search?q=org%3Aalphagov%20%22users.create_or_update_user%22&type=code
[2]: https://github.com/alphagov/support/pull/1244